requirements.txt
websockets>=10.0
aiohttp>=3.9.0
python-dotenv>=1.2.1
